requirements.txt
lxml>=3.7.2
openpyxl>=2.4.8
plac>=0.9.6
python-dateutil>=2.7.2
PyYAML>=3.12